# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /registration:
    post:
      summary: Endpoint for registering user
      operationId: registration
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        '200':
          description: Successful registration
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      summary: Endpoint to log user into system
      operationId: login
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: Successful login   
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/SuccessResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /get_profile:
    get:
      summary: Endpoint to get user profile
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/GetProfileSuccessResponse"
        '403':
          description: Forbidden Code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /update_profile:
    put:
      summary: Endpoint to get user profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      parameters:
        - name: phone_number
          in: query
          schema:
            type: string
        - name: full_name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful update profile
        '403':
          description: Forbidden Code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: Existing data already exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

   
components:
  securitySchemes:
      bearerAuth:           
        type: http
        scheme: bearer
        bearerFormat: JWT
  schemas:
    HelloResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    RegisterRequest:
      type: object
      required:
        - phone_number
        - password
        - full_name
      properties:
        phone_number:
          type: string
        full_name:
          type: string  
        password:
          type: string
    LoginRequest:
      type: object
      required:
        - phone_number
        - password
      properties:
        phone_number:
          type: string
        password:
          type: string
    SuccessResponse:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: integer
    GetProfileSuccessResponse:
      type: object
      required:
        - phone_number
        - full_name
      properties:
        phone_number:
          type: string
        full_name:
          type: string
          
security:
  - bearerAuth: []   